Moved the VisualStateSwitcher to its own file in preparation for required methods.
